refactor NavItem
