+++ aries-core/src/js/components/core/NavigationMenu/NavItem/ItemLabel.tsx
@@ -1,0 +1,15 @@
+import { Box, Text } from 'grommet';
+
+interface ItemLabelProps {
+  icon: React.ReactNode;
+  label: string;
+}
+
+export const ItemLabel = ({ icon, label }: ItemLabelProps) => {
+  return (
+    <Box direction="row" align="center" gap="xsmall" flex>
+      {icon}
+      <Text color="text-strong">{label}</Text>
+    </Box>
+  );
+};
