+++ aries-core/src/js/components/core/NavigationMenu/NavItem/index.ts
@@ -1,0 +1,1 @@
+export * from './NavItem';
